+++ src/api/rcos/discord_associations/project/create_project_channel.rs
@@ -9,25 +9,25 @@ use crate::error::TelescopeError;
 #[derive(GraphQLQuery)]
 #[graphql(
     schema_path = "graphql/rcos/schema.json",
-    query_path = "graphql/rcos/discord_assoications/small_group/create_small_group_channel.graphql"
+    query_path = "graphql/rcos/discord_associations/project/create_project_channel.graphql"
 )]
-pub struct CreateOneSmallGroupChannel;
+pub struct CreateOneProjectChannel;
 
-impl CreateOneSmallGroupChannel {
+impl CreateOneProjectChannel {
     pub async fn execute(
-        small_group_id: i64,
+        project_id: i64,
         channel_id: String,
         kind: channel_type,
     ) -> Result<Option<String>, TelescopeError> {
-        send_query::<Self>(create_one_small_group_channel::Variables {
-            small_group_id,
+        send_query::<Self>(create_one_project_channel::Variables {
+            project_id,
             channel_id,
             kind,
         })
         .await
         .map(|response| {
             response
-                .insert_small_group_channels_one
+                .insert_project_channels_one
                 .map(|obj| obj.channel_id)
         })
     }
