+++ src/api/rcos/discord_associations/project/project_info.rs
@@ -1,5 +1,5 @@
-use crate::api::rcos::discord_assoications::project::project_info::curr_projects::ResponseData;
-use crate::api::rcos::discord_assoications::project::project_info::curr_projects::Variables;
+use crate::api::rcos::discord_associations::project::project_info::curr_projects::ResponseData;
+use crate::api::rcos::discord_associations::project::project_info::curr_projects::Variables;
 use crate::api::rcos::{prelude::*, search_strings::resolve_search_string, send_query};
 use crate::error::TelescopeError;
 use chrono::Utc;
@@ -10,7 +10,7 @@ const PER_PAGE: u32 = 20;
 #[derive(GraphQLQuery)]
 #[graphql(
     schema_path = "graphql/rcos/schema.json",
-    query_path = "graphql/rcos/discord_assoications/project/projects.graphql",
+    query_path = "graphql/rcos/discord_associations/project/projects.graphql",
     response_derives = "Debug,Clone,Serialize"
 )]
 pub struct CurrProjects;
