+++ src/api/rcos/discord_associations/small_group/small_group_info.rs
@@ -1,5 +1,5 @@
-use crate::api::rcos::discord_assoications::small_group::small_group_info::curr_small_groups::ResponseData;
-use crate::api::rcos::discord_assoications::small_group::small_group_info::curr_small_groups::Variables;
+use crate::api::rcos::discord_associations::small_group::small_group_info::curr_small_groups::ResponseData;
+use crate::api::rcos::discord_associations::small_group::small_group_info::curr_small_groups::Variables;
 use crate::api::rcos::{prelude::*, search_strings::resolve_search_string, send_query};
 use crate::error::TelescopeError;
 use chrono::Utc;
@@ -11,7 +11,7 @@ const PER_PAGE: u32 = 20;
 #[derive(GraphQLQuery)]
 #[graphql(
     schema_path = "graphql/rcos/schema.json",
-    query_path = "graphql/rcos/discord_assoications/small_group/small_groups.graphql",
+    query_path = "graphql/rcos/discord_associations/small_group/small_groups.graphql",
     response_derives = "Debug,Clone,Serialize"
 )]
 pub struct CurrSmallGroups;
